requirements.txt
anyio==4.9.0
ariadne==0.26.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
deepmerge==2.0
dnspython==2.7.0
firebase_admin==7.1.0
Flask==3.1.1
flask-cors==6.0.1
Flask-PyMongo==3.0.1
flask-redis==0.4.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphql-core==3.2.5
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
msgpack==1.1.1
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.13.2
PySocks==1.7.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.4
rsa==4.9.1
selenium==4.21.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.42
starlette==0.47.2
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
